Make `expires_in` u32 and (on regenerate) not default to the same as last time

+++ crates/handlers/src/admin/v1/personal_sessions/add.rs
@@ -68,7 +68,7 @@ pub struct Request {
 
     /// Token expiry time in seconds.
     /// If not set, the token won't expire.
-    expires_in: Option<u64>,
+    expires_in: Option<u32>,
 }
 
 pub fn doc(operation: TransformOperation) -> TransformOperation {
